sprint15 reformat code from const
