+++ app/src/main/java/com/example/playlistmaker/ITunesApi.kt
@@ -1,5 +1,6 @@
 package com.example.playlistmaker
 
+import com.example.playlistmaker.domain.models.TrackResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
