+++ app/src/main/java/com/example/playlistmaker/TrackAdapter.kt
@@ -8,13 +8,13 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.playlistmaker.ui.AudioplayerActivity
-import com.example.playlistmaker.ui.KEY_TRACK_ID
+import com.example.playlistmaker.utils.consts.KEY_TRACK_ID
+import com.example.playlistmaker.utils.createJsonFromTrack
 
 open class TrackAdapter : RecyclerView.Adapter<PlaylistViewHolder>() {
     companion object {
         private const val CLICK_DEBOUNCE_DELAY = 1000L
     }
-
 
     var listTrack = mutableListOf<Track>()
     private var isClicked = true
