--- app/src/main/java/com/example/playlistmaker/TrackResponse.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker
-
-import com.example.playlistmaker.domain.models.Track
-
-data class TrackResponse(
-    val resultCount: Int,
-    val results: List<Track>
-)
