+++ app/src/main/java/com/example/playlistmaker/domain/models/TrackResponse.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.domain.models
+
+data class TrackResponse(
+    val resultCount: Int,
+    val results: List<Track>
+)
