+++ app/src/main/java/com/example/playlistmaker/ui/AudioplayerActivity.kt
@@ -11,8 +11,8 @@ import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.R
-import com.example.playlistmaker.createJsonFromTrack
-import com.example.playlistmaker.createTrackFromJson
+import com.example.playlistmaker.utils.createJsonFromTrack
+import com.example.playlistmaker.utils.createTrackFromJson
 import com.example.playlistmaker.creator.Creator
 import com.example.playlistmaker.dpToPx
 import com.example.playlistmaker.utils.consts.KEY_TRACK_ID
@@ -20,8 +20,6 @@ import java.text.SimpleDateFormat
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 import java.util.*
-
-
 
 class AudioplayerActivity : AppCompatActivity() {
