+++ app/src/main/java/com/example/playlistmaker/utils/consts/Consts.kt
@@ -1,3 +1,6 @@
 package com.example.playlistmaker.utils.consts
 
 const val KEY_TRACK_ID = "key_track_id"
+
+const val SEARCH_HISTORY_PLAYLIST = "search_history_playlist"
+const val SEARCH_HISTORY_NEW_TRACK = "search_history_new_track"
